Administration - Course List Json (#8)
* Create Maven Project

* Remove Bin

* Import Gson

Imports Gson for Json actions. Beginning of administration side of adding courses to a courselist.

* GSON Working

Got Reading/Writing Working for GSON JSON files.

* Update CourseList.java

* Finished Adding Courses Admin Task

Wrapped up the adding courses for now. We can always refactor where needed but this works as is and has a test constructor setup.

* Update CourseList.java

Removed unused imports